ref : replace image_error_widget.dart with image_placeholder_widget.dart
